requirements.txt
google-generativeai
google-cloud-vision
google-cloud-documentai
opencv-python
numpy
pytesseract
Pillow
ultralytics
torch
torchvision
python-dotenv
easyocr
pdf2image
